+++ cli/Valet/Memcache.php
@@ -4,63 +4,55 @@ namespace Valet;
 
 class Memcache
 {
-    public $brew;
-    public $cli;
-    public $pecl;
+    /** @var PhpExtension */
+    protected $phpExtension;
 
     /**
-     * Memcached.
-     *
-     * @param  Brew $brew
-     * @param  CommandLine $cli
-     * @param  Pecl $pecl
+     * @param PhpExtension $phpExtension
      */
-    public function __construct(Brew $brew, CommandLine $cli, Pecl $pecl)
-    {
-        $this->cli = $cli;
-        $this->brew = $brew;
-        $this->pecl = $pecl;
+    public function __construct(
+        PhpExtension $phpExtension
+    ) {
+        $this->phpExtension = $phpExtension;
     }
 
     /**
-     * Install service.
-     *
+     * Install memcache.
+     * @param $phpVersion
      * @return bool
      */
-    public function install()
+    public function install($phpVersion)
     {
-        $restart = false;
-        if ($this->brew->installed('libmemcached')) {
-            info('[libmemcached] (brew) already installed');
-        } else {
-            $restart = true;
-            $this->brew->ensureInstalled('libmemcached');
-            info('[libmemcached] Successfully installed');
+        if (!$this->phpExtension->isInstalled(PhpExtension::MEMCACHE_EXTENSION, $phpVersion)) {
+            return $this->phpExtension->installExtension(
+                PhpExtension::MEMCACHE_EXTENSION,
+                $phpVersion
+            );
         }
-        info('[memcached] Installing');
-        $peclInstalled = $this->pecl->installExtension('memcached');
-        if ($restart || $peclInstalled) {
-            return true;
-        }
+
+        info("[EXTENSION] Memcache extension is already installed!");
+
         return false;
     }
 
     /**
-     * Uninstall memcached.
-     *
+     * Uninstall memcache.
+     * @param $phpVersion
+     * @param $phpIniConfigPath
      * @return bool
      */
-    public function uninstall()
+    public function uninstall($phpVersion, $phpIniConfigPath)
     {
-        info('[memcached] Uninstalling');
-        $removed = $this->pecl->disableExtension('memcached');
-        if ($removed) {
-            info('[memcached] Successfully uninstalled');
-        } else {
-            info('[memcached] was already uninstalled');
+        if ($this->phpExtension->isInstalled(PhpExtension::MEMCACHE_EXTENSION, $phpVersion)) {
+            return $this->phpExtension->uninstallExtension(
+                PhpExtension::MEMCACHE_EXTENSION,
+                $phpVersion,
+                $phpIniConfigPath
+            );
         }
-        $this->brew->ensureUninstalled('libmemcached');
-        info('[libmemcached] Successfully uninstalled');
-        return true;
+
+        info("[EXTENSION] Memcache extension is already uninstalled!");
+
+        return false;
     }
 }