+++ cli/drivers/BedrockValetDriver.php
@@ -49,7 +49,7 @@ class BedrockValetDriver extends BasicValetDriver
     {
         $this->loadServerEnvironmentVariables($sitePath, $siteName);
 
-        $_SERVER['PHP_SELF'] = $uri;
+        $_SERVER['PHP_SELF']    = $uri;
         $_SERVER['SERVER_ADDR'] = '127.0.0.1';
         $_SERVER['SERVER_NAME'] = $_SERVER['HTTP_HOST'];