+++ CHANGELOG.md
@@ -4,7 +4,23 @@ All notable changes to valet-plus will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/weprovide/valet-plus/compare/2.3.1...2.x)
+## [2.4.0](https://github.com/weprovide/valet-plus/compare/2.3.1...2.4.0)
+### Fixed
+- [#612] Fixed Bedrock driver
+- [#522] Use SHA256 for self signed certificates
+- [#608] Update xdebug to support PHP 8.1
+
+### Added
+- Support for PHP 8.1
+- Support for PHP 8.2
+
+### Changed
+- Replaced valet-php tap by shivammathur php tap
+- Replaced pecl extensions by shivammathur php extensions
+
+### Removed
+- Support for PHP <7.1
+
 
 ## [2.3.1](https://github.com/weprovide/valet-plus/compare/2.3.0...2.3.1)
 ### Fixed